ULTIMAPRUEBA/TascaFinal/src/es/iesmz/ed/algoritmes/SudokuTest.java
package es.iesmz.ed.algoritmes;

import org.junit.jupiter.api.Test;
import static org.junit.jupiter.api.Assertions.assertEquals;

class SudokuTest {

    @Test
    public void testSudokuCorrecto() {
        int[][] sudokuCorrecto = {
                {5, 3, 4, 6, 7, 8, 9, 1, 2},
                {6, 7, 2, 1, 9, 5, 3, 4, 8},
                {1, 9, 8, 3, 4, 2, 5, 6, 7},
                {8, 5, 9, 7, 6, 1, 4, 2, 3},
                {4, 2, 6, 8, 5, 3, 7, 9, 1},
                {7, 1, 3, 9, 2, 4, 8, 5, 6},
                {9, 6, 1, 5, 3, 7, 2, 8, 4},
                {2, 8, 7, 4, 1, 9, 6, 3, 5},
                {3, 4, 5, 2, 8, 6, 1, 7, 9}
        };
        assertEquals(true, Sudoku.esSudokuCorrecto(sudokuCorrecto));
    }
    @Test
    public void testSudokuIncorrecto() {
        int[][] sudokuIncorrecto = {
                {5, 3, 4, 6, 7, 8, 9, 1, 2},
                {6, 7, 2, 1, 9, 5, 3, 4, 8},
                {1, 9, 8, 3, 4, 2, 5, 6, 7},
                {8, 5, 9, 7, 6, 1, 4, 2, 3},
                {4, 2, 6, 8, 5, 9, 7, 9, 1},
                {7, 1, 3, 9, 2, 4, 8, 5, 6},
                {9, 6, 1, 5, 3, 7, 2, 8, 4},
                {2, 8, 7, 4, 1, 9, 6, 3, 5},
                {7, 4, 5, 2, 5, 6, 1, 7, 1}
                // He agregado un número incorrecto en el principio de la última
                // fila en medio y el final y en el medio
        };
        assertEquals(false, Sudoku.esSudokuCorrecto(sudokuIncorrecto));
    }

}